+++ src/modules/page/WellnessClinicHelper.js
@@ -1,6 +1,7 @@
 const WellnessClinicHelper = {
     cmds: 'wellness_clinic',
     staff: false,
+    localKeys: ['hw_wellness_goal'],
     init: function() {
         const url = window.location.href;
 
@@ -133,4 +134,3 @@ const WellnessClinicHelper = {
         }, 150);
     }
 }
-